Trait compiler_builtins::float::Float

pub trait Float: Copy + PartialEq + PartialOrd + AddAssign + MulAssign + Add<Output = Self> + Sub<Output = Self> + Div<Output = Self> + Rem<Output = Self> {
    type Int: Int;

    const ZERO: Self;
    const ONE: Self;
    const BITS: u32;
    const SIGNIFICAND_BITS: u32;
    const EXPONENT_BITS: u32;
    const EXPONENT_MAX: u32;
    const EXPONENT_BIAS: u32;
    const SIGN_MASK: Self::Int;
    const SIGNIFICAND_MASK: Self::Int;
    const IMPLICIT_BIT: Self::Int;
    const EXPONENT_MASK: Self::Int;

    fn repr(self) -> Self::Int;

    fn from_repr(a: Self::Int) -> Self;

    fn from_parts(
        sign: bool, 
        exponent: Self::Int, 
        significand: Self::Int
    ) -> Self;

    fn normalize(significand: Self::Int) -> (i32, Self::Int);
}

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

Trait for some basic operations on floats

Associated Types

type Int: Int

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

A uint of the same with as the float

Associated Constants

const ZERO: Self

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

const ONE: Self

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

const BITS: u32

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

The bitwidth of the float type

const SIGNIFICAND_BITS: u32

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

The bitwidth of the significand

const EXPONENT_BITS: u32

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

EXPONENT_BITS: u32 = <Self>::BITS - <Self>::SIGNIFICAND_BITS - 1

The bitwidth of the exponent

const EXPONENT_MAX: u32

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

EXPONENT_MAX: u32 = (1 << <Self>::EXPONENT_BITS) - 1

The maximum value of the exponent

const EXPONENT_BIAS: u32

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

EXPONENT_BIAS: u32 = <Self>::EXPONENT_MAX >> 1

The exponent bias value

const SIGN_MASK: Self::Int

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

A mask for the sign bit

const SIGNIFICAND_MASK: Self::Int

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

A mask for the significand

const IMPLICIT_BIT: Self::Int

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

const EXPONENT_MASK: Self::Int

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

A mask for the exponent

Required Methods

fn repr(self) -> Self::Int

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

Returns self transmuted to Self::Int

fn from_repr(a: Self::Int) -> Self

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

Returns a Self::Int transmuted back to Self

fn from_parts(sign: bool, exponent: Self::Int, significand: Self::Int) -> Self

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

Constructs a Self from its parts. Inputs are treated as bits and shifted into position.

fn normalize(significand: Self::Int) -> (i32, Self::Int)

🔬 This is a nightly-only experimental API. (compiler_builtins_lib)

Compiler builtins. Will never become stable.

Returns (normalized exponent, normalized significand)