Module compiler_builtins::int

Re-exports

pub use self::leading_zeros::__clzsi2;

pub use self::leading_zeros::__clzsi2;

Modules

addsub

leading_zeros

mul

sdiv

shift

specialized_div_rem 🔒

This specialized_div_rem module is originally from version 1.0.0 of the specialized-div-rem crate. Note that for loops with ranges are not used in this module, since unoptimized compilation may generate references to memcpy.

udiv

Macros

cast_into 🔒

impl_d_int 🔒

impl_h_int 🔒

int_impl 🔒

int_impl_common 🔒

Traits

CastInto 🔒

Trait to express (possibly lossy) casting of integers

DInt 🔒

Trait for integers twice the bit width of another integer. This is implemented for all primitives except for u8, because there is not a smaller primitive.

HInt 🔒

Trait for integers half the bit width of another integer. This is implemented for all primitives except for u128, because it there is not a larger primitive.

Int 🔒

Trait for some basic operations on integers