compiler_builtins::int

Module leading_zeros

Functions

__clzdi2

Returns the number of leading binary zeros in x

__clzsi2

Returns the number of leading binary zeros in x

__clzti2

Returns the number of leading binary zeros in x