Trait compiler_builtins::int::DInt

pub(crate) trait DInt: Int {
    type H: HInt<D = Self> + Int;

    // Required methods
    fn lo(self) -> Self::H;
    fn hi(self) -> Self::H;
    fn lo_hi(self) -> (Self::H, Self::H);
    fn from_lo_hi(lo: Self::H, hi: Self::H) -> Self;
}

Trait for integers twice the bit width of another integer. This is implemented for all primitives except for u8, because there is not a smaller primitive.

Required Associated Types

type H: HInt<D = Self> + Int

Integer that is half the bit width of the integer this trait is implemented for

Required Methods

fn lo(self) -> Self::H

Returns the low half of self

fn hi(self) -> Self::H

Returns the high half of self

fn lo_hi(self) -> (Self::H, Self::H)

Returns the low and high halves of self as a tuple

fn from_lo_hi(lo: Self::H, hi: Self::H) -> Self

Constructs an integer using lower and higher half parts