Trait compiler_builtins::int::HInt

pub(crate) trait HInt: Int {
    type D: DInt<H = Self> + Int;

    // Required methods
    fn widen(self) -> Self::D;
    fn zero_widen(self) -> Self::D;
    fn widen_hi(self) -> Self::D;
    fn zero_widen_mul(self, rhs: Self) -> Self::D;
    fn widen_mul(self, rhs: Self) -> Self::D;
}

Trait for integers half the bit width of another integer. This is implemented for all primitives except for u128, because it there is not a larger primitive.

Required Associated Types

type D: DInt<H = Self> + Int

Integer that is double the bit width of the integer this trait is implemented for

Required Methods

fn widen(self) -> Self::D

Widens (using default extension) the integer to have double bit width

fn zero_widen(self) -> Self::D

Widens (zero extension only) the integer to have double bit width. This is needed to get around problems with associated type bounds (such as Int<Othersign: DInt>) being unstable

fn widen_hi(self) -> Self::D

Widens the integer to have double bit width and shifts the integer into the higher bits

fn zero_widen_mul(self, rhs: Self) -> Self::D

Widening multiplication with zero widening. This cannot overflow.

fn widen_mul(self, rhs: Self) -> Self::D

Widening multiplication. This cannot overflow.