compiler_builtins::math

Module full_availability

This set of functions is well tested in libm and known to provide similar performance to system libm, as well as the same or better accuracy.

Functions

cbrt

cbrtf

ceil

ceilf

ceilf16

ceilf128

copysign

copysignf

copysignf16

copysignf128

fabs

fabsf

fabsf16

fabsf128

fdim

fdimf

fdimf16

fdimf128

floor

floorf

floorf16

floorf128

fma

fmaf

fmaf128

fmax

fmaxf

fmaxf16

fmaxf128

fmaximum

fmaximumf

fmaximumf16

fmaximumf128

fmin

fminf

fminf16

fminf128

fminimum

fminimumf

fminimumf16

fminimumf128

fmod

fmodf

fmodf16

fmodf128

rint

rintf

rintf16

rintf128

round

roundeven

roundevenf

roundevenf16

roundevenf128

roundf

roundf16

roundf128

sqrt

sqrtf

sqrtf16

sqrtf128

trunc

truncf

truncf16

truncf128