Trait compiler_builtins::float::Float

pub(crate) trait Float: Copy + Debug + PartialEq + PartialOrd + AddAssign + MulAssign + Add<Output = Self> + Sub<Output = Self> + Div<Output = Self> + Rem<Output = Self> {
    type Int: Int;
    type SignedInt: Int;
    type ExpInt: Int;

    const ZERO: Self;
    const ONE: Self;
    const BITS: u32;
    const SIGNIFICAND_BITS: u32;
    const SIGN_MASK: Self::Int;
    const SIGNIFICAND_MASK: Self::Int;
    const IMPLICIT_BIT: Self::Int;
    const EXPONENT_MASK: Self::Int;
    const EXPONENT_BITS: u32 = _;
    const EXPONENT_MAX: u32 = _;
    const EXPONENT_BIAS: u32 = _;

    // Required methods
    fn repr(self) -> Self::Int;
    fn signed_repr(self) -> Self::SignedInt;
    fn eq_repr(self, rhs: Self) -> bool;
    fn sign(self) -> bool;
    fn exp(self) -> Self::ExpInt;
    fn frac(self) -> Self::Int;
    fn imp_frac(self) -> Self::Int;
    fn from_repr(a: Self::Int) -> Self;
    fn from_parts(
        sign: bool,
        exponent: Self::Int,
        significand: Self::Int
    ) -> Self;
    fn normalize(significand: Self::Int) -> (i32, Self::Int);
    fn is_subnormal(self) -> bool;

}

Trait for some basic operations on floats

Required Associated Types

type Int: Int

A uint of the same width as the float

type SignedInt: Int

A int of the same width as the float

type ExpInt: Int

An int capable of containing the exponent bits plus a sign bit. This is signed.

Required Associated Constants

const ZERO: Self

const ONE: Self

const BITS: u32

The bitwidth of the float type

const SIGNIFICAND_BITS: u32

The bitwidth of the significand

const SIGN_MASK: Self::Int

A mask for the sign bit

const SIGNIFICAND_MASK: Self::Int

A mask for the significand

const IMPLICIT_BIT: Self::Int

const EXPONENT_MASK: Self::Int

A mask for the exponent

Provided Associated Constants

const EXPONENT_BITS: u32 = _

The bitwidth of the exponent

const EXPONENT_MAX: u32 = _

The maximum value of the exponent

const EXPONENT_BIAS: u32 = _

The exponent bias value

Required Methods

fn repr(self) -> Self::Int

Returns self transmuted to Self::Int

fn signed_repr(self) -> Self::SignedInt

Returns self transmuted to Self::SignedInt

fn eq_repr(self, rhs: Self) -> bool

Checks if two floats have the same bit representation. Except for NaNs! NaN can be represented in multiple different ways. This method returns true if two NaNs are compared.

fn sign(self) -> bool

Returns the sign bit

fn exp(self) -> Self::ExpInt

Returns the exponent with bias

fn frac(self) -> Self::Int

Returns the significand with no implicit bit (or the “fractional” part)

fn imp_frac(self) -> Self::Int

Returns the significand with implicit bit

fn from_repr(a: Self::Int) -> Self

Returns a Self::Int transmuted back to Self

fn from_parts(sign: bool, exponent: Self::Int, significand: Self::Int) -> Self

Constructs a Self from its parts. Inputs are treated as bits and shifted into position.

fn normalize(significand: Self::Int) -> (i32, Self::Int)

Returns (normalized exponent, normalized significand)

fn is_subnormal(self) -> bool

Returns if self is subnormal